misc/HtmlTestSuite/results/test-0869.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>FileReader NotReadableError Test</title>
<link href="http://www.intel.com" rel="author" title="Intel"/>
<link href="https://w3c.github.io/FileAPI/#dfn-error-codes" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
</head><body><form name="upload">
  <input id="fileChooser" type="file"/><br/><input id="start" type="button" value="start"/>
</form>

<div>
  <p>Test steps:</p>
  <ol>
    <li>Download the <a href="support/file_test1.txt">file</a>.</li>
    <li>Select the file in the file inputbox.</li>
    <li>Delete the file's readable permission.</li>
    <li>Click the 'start' button.</li>
  </ol>
</div>

<script>

  const fileChooser = document.querySelector('#fileChooser');

  setup({explicit_done: true});
  setup({explicit_timeout: true});

  on_event(document.querySelector('#start'), 'click', () =&gt; {
    async_test(t =&gt; {
      const reader = new FileReader();
      reader.readAsArrayBuffer(fileChooser.files[0]);
      reader.onloadend = t.step_func_done(event =&gt; {
        assert_equals(event.target.readyState, FileReader.DONE);
        assert_equals(reader.error.name, "NotReadableError");
      });
    }, 'FileReader.error should be NotReadableError if the file is unreadable');
    done();
  });

</script>

</body></html>
